refactor(sharing): split code into multiple files
