+++ crates/plex-api/src/myplex/sharing/shareables.rs
@@ -1,0 +1,47 @@
+use crate::{Library, MyPlex, Server};
+
+pub enum ShareableServer<'a> {
+    MachineIdentifier(&'a str),
+    Server(&'a Server),
+}
+
+impl<'a> ShareableServer<'a> {
+    pub fn get_id(&self) -> &str {
+        match self {
+            Self::MachineIdentifier(id) => id,
+            Self::Server(srv) => srv.machine_identifier(),
+        }
+    }
+}
+
+pub enum ShareableLibrary<'a> {
+    Library(&'a Library),
+    LibraryId(&'a str),
+}
+
+impl<'a> ShareableLibrary<'a> {
+    pub fn get_id(&self) -> &str {
+        match self {
+            Self::Library(library) => library.id(),
+            Self::LibraryId(id) => id,
+        }
+    }
+}
+
+pub enum User<'a> {
+    Account(&'a MyPlex),
+    UsernameOrEmail(&'a str),
+}
+
+impl<'a> User<'a> {
+    pub fn get_identifier(&self) -> &str {
+        match self {
+            Self::Account(MyPlex {
+                account: Some(account),
+                ..
+            }) => &account.email,
+            Self::Account(_) => "",
+            Self::UsernameOrEmail(u) => u,
+        }
+    }
+}
